Move the generic code under a top-level c_analyzer package.
