+++ Lib/test/test_tools/test_c_analyzer/test_c_globals/test___main__.py
@@ -3,8 +3,8 @@ import unittest
 
 from .. import tool_imports_for_tests
 with tool_imports_for_tests():
-    from c_analyzer_common import SOURCE_DIRS, info
-    from c_analyzer_common.known import DATA_FILE as KNOWN_FILE
+    from c_analyzer.common import SOURCE_DIRS, info
+    from c_analyzer.common.known import DATA_FILE as KNOWN_FILE
     import c_globals as cg
     from c_globals.supported import IGNORED_FILE
     from c_globals.__main__ import cmd_check, cmd_show, parse_args, main
