+++ Lib/test/test_tools/test_c_analyzer/test_common/test_info.py
@@ -4,7 +4,7 @@ import unittest
 from ..util import PseudoStr, StrProxy, Object
 from .. import tool_imports_for_tests
 with tool_imports_for_tests():
-    from c_analyzer_common.info import (
+    from c_analyzer.common.info import (
             UNKNOWN,
             ID,
             normalize_vartype, Variable
