+++ Lib/test/test_tools/test_c_analyzer/test_common/test_known.py
@@ -4,8 +4,8 @@ import unittest
 
 from .. import tool_imports_for_tests
 with tool_imports_for_tests():
-    from c_analyzer_common.info import ID, Variable
-    from c_analyzer_common.known import from_file
+    from c_analyzer.common.info import Variable
+    from c_analyzer.common.known import from_file
 
 
 class FromFileTests(unittest.TestCase):
