+++ Lib/test/test_tools/test_c_analyzer/test_parser/test_declarations.py
@@ -3,7 +3,7 @@ import unittest
 
 from .. import tool_imports_for_tests
 with tool_imports_for_tests():
-    from c_parser.declarations import (
+    from c_analyzer.parser.declarations import (
         iter_global_declarations, iter_local_statements,
         parse_func, _parse_var, parse_compound,
         iter_variables,
