+++ Tools/c-analyzer/c_analyzer/common/__init__.py
@@ -2,10 +2,14 @@ import os.path
 import sys
 
 
-PKG_ROOT = os.path.dirname(__file__)
-DATA_DIR = os.path.dirname(PKG_ROOT)
-REPO_ROOT = os.path.dirname(
-        os.path.dirname(DATA_DIR))
+TOOL_ROOT = (
+        os.path.dirname(  # c-analyzer/
+            os.path.dirname(  # c_analyzer/
+                os.path.dirname(__file__))))  # common/
+DATA_DIR = TOOL_ROOT
+REPO_ROOT = (
+        os.path.dirname(  # ..
+            os.path.dirname(TOOL_ROOT)))  # Tools/
 
 SOURCE_DIRS = [os.path.join(REPO_ROOT, name) for name in [
         'Include',
