+++ Tools/c-analyzer/c_analyzer/common/show.py
