+++ Tools/c-analyzer/c_analyzer/common/util.py
