+++ Tools/c-analyzer/c_analyzer/parser/naive.py
@@ -1,6 +1,6 @@
 import re
 
-from c_analyzer_common.info import UNKNOWN, Variable
+from ..common.info import UNKNOWN, Variable
 
 from .preprocessor import _iter_clean_lines
